updates  ProductReview Service Updates

+++ libs/common/src/main/java/smr/shop/libs/common/dto/message/ProductReviewMessageModel.java
@@ -20,7 +20,7 @@ public class ProductReviewMessageModel implements BaseMessageModel {
 
     private Long productId;
 
-    private Byte rating;
+    private float rating;
 
     private String comment;
